Test Enhancements and Bug fixes

+++ test_pool/pcie/test_p012.c
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2016-2018, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2016-2018,2022 Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -36,7 +36,7 @@ payload (void)
   uint32_t index;
   uint32_t count;
   PERIPHERAL_IRQ_MAP *irq_map;
-  uint8_t status;
+  uint32_t status;
   uint32_t current_irq_pin;
   uint32_t next_irq_pin;
   uint64_t dev_bdf;
